Casting size_t to int in '6-size.c'

+++ 0x00-hello_world/6-size.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-	printf("Size of a char: %zd byte(s)\n", sizeof(char));
-	printf("Size of an int: %zd byte(s)\n", sizeof(int));
-	printf("Size of a long int: %zd byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %zd byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %zd byte(s)\n", sizeof(float));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", (int) sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(long long int));
+	printf("Size of a float: %d byte(s)\n", (int) sizeof(float));
 	return (0);
 }
